+++ .ansible/collections/ansible_collections/netapp/ontap/plugins/modules/na_ontap_fpolicy_event.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python
 
-# (c) 2021-2025, NetApp, Inc
+# (c) 2021-2026, NetApp, Inc
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 
 from __future__ import absolute_import, division, print_function
@@ -64,8 +64,31 @@ options:
     - Indicator if the volume operation required for the event. If not specified the default Value is false.
     type: bool
 
+  lambda_config:
+    description:
+      - Configuration parameters for AWS Lambda proxy functionality.
+      - These option and suboptions are only supported with REST.
+    type: dict
+    version_added: 23.6.0
+    suboptions:
+      function_name:
+        description:
+          - The name of the AWS Lambda function to invoke.
+        type: str
+        required: true
+      aws_region:
+        description:
+          - The name of the AWS region.
+        type: str
+        required: true
+      aws_profile:
+        description:
+          - The name of the AWS profile to use for authentication.
+        type: str
+
 notes:
-- Support check_mode.
+  - Support check_mode.
+  - Supports AWS Lambda proxy functionality when using REST. See the README file for examples.
 '''
 
 EXAMPLES = """
@@ -141,11 +164,15 @@ class NetAppOntapFpolicyEvent():
             protocol=dict(required=False, type='str', choices=['cifs', 'nfsv3', 'nfsv4']),
             volume_monitoring=dict(required=False, type='bool')
         ))
+        self.argument_spec.update(netapp_utils.na_ontap_lambda_argument_spec())
         self.module = AnsibleModule(
             argument_spec=self.argument_spec,
             supports_check_mode=True,
             required_together=[
-                ('protocol', 'file_operations')]
+                ('protocol', 'file_operations')],
+            required_if=[
+                ['use_lambda', True, ('lambda_config',)]
+            ],
         )
         self.na_helper = NetAppModule()
         self.parameters = self.na_helper.set_parameters(self.module.params)
@@ -155,6 +182,8 @@ class NetAppOntapFpolicyEvent():
         self.use_rest = self.rest_api.is_rest()
 
         if not self.use_rest:
+            if self.parameters.get('use_lambda'):
+                self.module.fail_json(msg="Error: AWS Lambda proxy for ONTAP APIs is only supported with REST.")
             if not netapp_utils.has_netapp_lib():
                 self.module.fail_json(msg=netapp_utils.netapp)
             else: