+++ .ansible/collections/ansible_collections/netapp/ontap/tests/unit/plugins/modules/test_na_ontap_fpolicy_status.py
@@ -1,4 +1,4 @@
-# (c) 2021, NetApp, Inc
+# (c) 2021-2026, NetApp, Inc
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 
 ''' unit test for ONTAP fpolicy status Ansible module '''
@@ -16,8 +16,8 @@ from ansible_collections.netapp.ontap.tests.unit.plugins.module_utils.ansible_mo
 from ansible_collections.netapp.ontap.plugins.modules.na_ontap_fpolicy_status \
     import NetAppOntapFpolicyStatus as my_module  # module under test
 
-if not netapp_utils.has_netapp_lib():
-    pytestmark = pytest.mark.skip('skipping as missing required netapp_lib')
+# if not netapp_utils.has_netapp_lib():
+#    pytestmark = pytest.mark.skip('skipping as missing required netapp_lib')
 
 
 class MockONTAPConnection():
@@ -145,6 +145,7 @@ def get_fpolicy_status_mock_object(cx_type='zapi', kind=None):
     return fpolicy_status_obj
 
 
+@pytest.mark.skipif(not netapp_utils.has_netapp_lib(), reason="skipping as missing required netapp_lib")
 def test_module_fail_when_required_args_missing(patch_ansible):
     ''' required arguments are reported as errors '''
     with pytest.raises(AnsibleFailJson) as exc:
@@ -153,6 +154,7 @@ def test_module_fail_when_required_args_missing(patch_ansible):
     print('Info: %s' % exc.value.args[0]['msg'])
 
 
+@pytest.mark.skipif(not netapp_utils.has_netapp_lib(), reason="skipping as missing required netapp_lib")
 def test_ensure_get_called(patch_ansible):
     ''' test get_fpolicy_policy_status for non-existent fPolicy'''
     args = dict(default_args())
@@ -165,6 +167,7 @@ def test_ensure_get_called(patch_ansible):
     assert my_obj.get_fpolicy_policy_status is not None
 
 
+@pytest.mark.skipif(not netapp_utils.has_netapp_lib(), reason="skipping as missing required netapp_lib")
 def test_rest_missing_arguments(patch_ansible):     # pylint: disable=redefined-outer-name,unused-argument
     ''' enable fpolicy '''
     args = dict(default_args())
@@ -176,6 +179,7 @@ def test_rest_missing_arguments(patch_ansible):     # pylint: disable=redefined-
     assert exc.value.args[0]['msg'] == msg
 
 
+@pytest.mark.skipif(not netapp_utils.has_netapp_lib(), reason="skipping as missing required netapp_lib")
 @patch('ansible_collections.netapp.ontap.plugins.modules.na_ontap_fpolicy_status.NetAppOntapFpolicyStatus.enable_fpolicy_policy')
 def test_successful_enable(self, patch_ansible):
     ''' Enable fPolicy and test idempotency '''
@@ -202,6 +206,7 @@ def test_successful_enable(self, patch_ansible):
     assert not exc.value.args[0]['changed']
 
 
+@pytest.mark.skipif(not netapp_utils.has_netapp_lib(), reason="skipping as missing required netapp_lib")
 @patch('ansible_collections.netapp.ontap.plugins.modules.na_ontap_fpolicy_status.NetAppOntapFpolicyStatus.disable_fpolicy_policy')
 def test_successful_disable(self, patch_ansible):
     ''' Disable fPolicy and test idempotency '''
@@ -230,6 +235,7 @@ def test_successful_disable(self, patch_ansible):
     assert not exc.value.args[0]['changed']
 
 
+@pytest.mark.skipif(not netapp_utils.has_netapp_lib(), reason="skipping as missing required netapp_lib")
 def test_if_all_methods_catch_exception(patch_ansible):
     args = dict(default_args())
     args['use_rest'] = 'never'