+++ plugins/modules/na_ontap_fpolicy_policy.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python
 
-# (c) 2021-2025, NetApp, Inc
+# (c) 2021-2026, NetApp, Inc
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 
 from __future__ import absolute_import, division, print_function
@@ -73,6 +73,30 @@ options:
     - User name for privileged access.
     type: str
 
+  lambda_config:
+    description:
+      - Configuration parameters for AWS Lambda proxy functionality.
+      - These option and suboptions are only supported with REST.
+    type: dict
+    version_added: 23.6.0
+    suboptions:
+      function_name:
+        description:
+          - The name of the AWS Lambda function to invoke.
+        type: str
+        required: true
+      aws_region:
+        description:
+          - The name of the AWS region.
+        type: str
+        required: true
+      aws_profile:
+        description:
+          - The name of the AWS profile to use for authentication.
+        type: str
+
+notes:
+  - Supports AWS Lambda proxy functionality when using REST. See the README file for examples.
 '''
 
 EXAMPLES = """
@@ -136,9 +160,13 @@ class NetAppOntapFpolicyPolicy():
             is_passthrough_read_enabled=dict(required=False, type='bool'),
             privileged_user_name=dict(required=False, type='str')
         ))
+        self.argument_spec.update(netapp_utils.na_ontap_lambda_argument_spec())
         self.module = AnsibleModule(
             argument_spec=self.argument_spec,
-            supports_check_mode=True
+            supports_check_mode=True,
+            required_if=[
+                ['use_lambda', True, ('lambda_config',)]
+            ],
         )
 
         self.na_helper = NetAppModule()
@@ -148,6 +176,8 @@ class NetAppOntapFpolicyPolicy():
         self.use_rest = self.rest_api.is_rest()
 
         if not self.use_rest:
+            if self.parameters.get('use_lambda'):
+                self.module.fail_json(msg="Error: AWS Lambda proxy for ONTAP APIs is only supported with REST.")
             if not netapp_utils.has_netapp_lib():
                 self.module.fail_json(msg=netapp_utils.netapp_lib_is_required())
             else:
@@ -327,11 +357,11 @@ class NetAppOntapFpolicyPolicy():
         """
         if self.use_rest:
             api = "/private/cli/vserver/fpolicy/policy"
-            body = {
+            query = {
                 'vserver': self.parameters['vserver'],
                 'policy-name': self.parameters['name']
             }
-            dummy, error = self.rest_api.delete(api, body)
+            dummy, error = self.rest_api.delete(api, params=query)
             if error:
                 self.module.fail_json(msg=error)