+++ plugins/modules/na_ontap_fpolicy_scope.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python
 
-# (c) 2021-2025, NetApp, Inc
+# (c) 2021-2026, NetApp, Inc
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 
 from __future__ import absolute_import, division, print_function
@@ -97,6 +97,30 @@ options:
     type: list
     elements: str
 
+  lambda_config:
+    description:
+      - Configuration parameters for AWS Lambda proxy functionality.
+      - These option and suboptions are only supported with REST.
+    type: dict
+    version_added: 23.6.0
+    suboptions:
+      function_name:
+        description:
+          - The name of the AWS Lambda function to invoke.
+        type: str
+        required: true
+      aws_region:
+        description:
+          - The name of the AWS region.
+        type: str
+        required: true
+      aws_profile:
+        description:
+          - The name of the AWS profile to use for authentication.
+        type: str
+
+notes:
+  - Supports AWS Lambda proxy functionality when using REST. See the README file for examples.
 '''
 
 EXAMPLES = """
@@ -164,9 +188,13 @@ class NetAppOntapFpolicyScope():
             volumes_to_exclude=dict(required=False, type='list', elements='str'),
             volumes_to_include=dict(required=False, type='list', elements='str')
         ))
+        self.argument_spec.update(netapp_utils.na_ontap_lambda_argument_spec())
         self.module = AnsibleModule(
             argument_spec=self.argument_spec,
-            supports_check_mode=True
+            supports_check_mode=True,
+            required_if=[
+                ['use_lambda', True, ('lambda_config',)]
+            ],
         )
 
         self.na_helper = NetAppModule()
@@ -176,6 +204,8 @@ class NetAppOntapFpolicyScope():
         self.use_rest = self.rest_api.is_rest()
 
         if not self.use_rest:
+            if self.parameters.get('use_lambda'):
+                self.module.fail_json(msg="Error: AWS Lambda proxy for ONTAP APIs is only supported with REST.")
             if not netapp_utils.has_netapp_lib():
                 self.module.fail_json(msg=netapp_utils.netapp_lib_is_required())
             else:
@@ -280,7 +310,8 @@ is-file-extension-check-on-directories-enabled,is-monitoring-of-objects-with-no-
             for parameter in (
                 'export_policies_to_exclude', 'export_policies_to_include', 'export_policies_to_include', 'file_extensions_to_exclude',
                 'file_extensions_to_include', 'shares_to_exclude', 'shares_to_include', 'volumes_to_exclude', 'volumes_to_include',
-                'is-file-extension-check-on-directories-enabled', 'is-monitoring-of-objects-with-no-extension-enabled'
+                'is-file-extension-check-on-directories-enabled', 'is-monitoring-of-objects-with-no-extension-enabled',
+                'check_extensions_on_directories'
             ):
                 if parameter in self.parameters:
                     body[parameter.replace('_', '-')] = self.parameters[parameter]
@@ -462,11 +493,11 @@ is-file-extension-check-on-directories-enabled,is-monitoring-of-objects-with-no-
 
         if self.use_rest:
             api = "/private/cli/vserver/fpolicy/policy/scope"
-            body = {
+            query = {
                 'vserver': self.parameters['vserver'],
                 'policy-name': self.parameters['name']
             }
-            dummy, error = self.rest_api.delete(api, body)
+            dummy, error = self.rest_api.delete(api, params=query)
             if error:
                 self.module.fail_json(msg=error)
         else: